Fixed indexing error in formatting
An error was occurring when attempting to add a second set of formatting
on text from the end of the first formatting to the end of the message.
This was due to an indexing error in a while loop: changed it to while
i<=text.length() and that fixed it.