Multiple msgs/notifications are combined
If there are multiple unread messages in the inbox when a new message
comes in, the notification will display the proper number of unread
messages.
If a new notification comes in when there is an existing notification in
the status tray, it will replace the existing notification (with an
updated count).