Messages sent from compose view added to db
Messages sent from the compose screen are now also added to the sent
messages database. Still need to see if this will work when a message is
sent from the compose screen to a contact with whom a prior conversation
thread does not exist. If not, a new thread and thread_id will have to
be created.