Temporary fix for back from conversation bug
Implemented a very bad fix for the bug where pressing back from the
conversation view closes the app. This is not a permanent solution - at
some point I need an action bar with a top-left corner button that does
this (goes back to the conversation no matter what) but for ease of
development purposes in the meantime i've put it here. I'm not sure what
the back button is actually supposed to do in this situation but I need
to figure it out.