Conversation thread is updated when msg is sent
When in the conversation view, sent messages are added to the thread
upon successful send. This is done through a receiver that is
initialized in onResume and detached in onPause so if you navigate out
of the conversation view before your message is sent it doesn't trigger.
Unfortunately, the message is added to the sent databse by the same
receiver, so if the user leaves the conversation view the message
doesn't get added to the database. Will fix this later.