Bugfix: going back from convo before msg send
Fixed bug from ae14e58f0cf08114205c4ba2bb8993788c0c7ae8.  Previously,
going back to the inbox from a conversation after pressing send but
before the message actually sent was causing the message to not be added
to the database. Now it gets added, but unfortunately the inbox
refreshes BEFORE the message is sent so it doesn't show the recently
sent message, which messes up the index of the list items.