Bugfix: going back from convo before msg send 2
(sort of) fixed the bug from ab62d38a0bbc9df1081e81d9f7b387c81efea6b8.
It's a terrible fix that should be replaced before even a beta is
released but it works for the meantime. It simply stops the user from
navigating back to the inbox while a message is trying to send.